refactor: unit test sync with new architecture
