+++ test/unit/NestedAssetBatcher.unit.ts
@@ -1,9 +1,10 @@
 import { LoadFixtureFunction } from "../types";
-import { factoryAndOperatorsFixture, FactoryAndOperatorsFixture } from "../shared/fixtures";
-import { createFixtureLoader, describeWithoutFork, expect, provider } from "../shared/provider";
+import { createFixtureLoader, describeWithoutFork, expect, provider } from "../shared/helpers/provider";
 import { BigNumber } from "ethers";
 import { appendDecimals, getExpectedFees } from "../helpers";
 import * as utils from "../../scripts/utils";
+import { factoryAndOperatorsFixture } from "../shared/helpers/fixtures/factoryAndOperatorsFixture";
+import { FactoryAndOperatorsFixture } from "../shared/types/FactoryAndOperatorsFixture";
 
 let loadFixture: LoadFixtureFunction;
 
@@ -29,15 +30,15 @@ describeWithoutFork("NestedAssetBatcher", () => {
         beforeEach("Create NFT (id 1)", async () => {
             // create nft 1 with UNI and KNC from DAI (use the base amounts)
             let orders: utils.OrderStruct[] = utils.getUniAndKncWithDaiOrders(context, baseUniBought, baseKncBought);
-            await context.nestedFactory
-                .connect(context.user1)
+            await context.nested.nestedFactory
+                .connect(context.wallets.user1)
                 .create(0, [
-                    { inputToken: context.mockDAI.address, amount: baseTotalToSpend, orders, fromReserve: true },
+                    { inputToken: context.tokens.mockDAI.address, amount: baseTotalToSpend, orders, fromReserve: true },
                 ]);
         });
 
         it("get all ids", async () => {
-            expect(await (await context.nestedAssetBatcher.getIds(context.user1.address)).toString()).to.equal(
+            expect(await (await context.nested.nestedAssetBatcher.getIds(context.wallets.user1.address)).toString()).to.equal(
                 [BigNumber.from(1)].toString(),
             );
         });
@@ -47,26 +48,26 @@ describeWithoutFork("NestedAssetBatcher", () => {
                 {
                     id: BigNumber.from(1),
                     assets: [
-                        { token: context.mockUNI.address, qty: baseUniBought },
-                        { token: context.mockKNC.address, qty: baseKncBought },
+                        { token: context.tokens.mockUNI.address, qty: baseUniBought },
+                        { token: context.tokens.mockKNC.address, qty: baseKncBought },
                     ],
                 },
             ];
 
-            const nfts = await context.nestedAssetBatcher.getNfts(context.user1.address);
+            const nfts = await context.nested.nestedAssetBatcher.getNfts(context.wallets.user1.address);
 
             expect(JSON.stringify(utils.cleanResult(nfts))).to.equal(JSON.stringify(utils.cleanResult(expectedNfts)));
         });
 
         it("require and get TokenHoldings", async () => {
-            await expect(context.nestedAssetBatcher.requireTokenHoldings(2)).to.be.revertedWith("NAB: NEVER_CREATED");
-            await expect(context.nestedAssetBatcher.requireTokenHoldings(1)).to.not.be.reverted;
+            await expect(context.nested.nestedAssetBatcher.requireTokenHoldings(2)).to.be.revertedWith("NAB: NEVER_CREATED");
+            await expect(context.nested.nestedAssetBatcher.requireTokenHoldings(1)).to.not.be.reverted;
 
             let orders: utils.OrderStruct[] = utils.getUsdcWithUniAndKncOrders(context, baseUniBought, baseKncBought);
-            await context.nestedFactory.connect(context.user1).destroy(1, context.mockUSDC.address, orders);
+            await context.nested.nestedFactory.connect(context.wallets.user1).destroy(1, context.tokens.mockUSDC.address, orders);
 
             // Not reverting after burn
-            await expect(context.nestedAssetBatcher.requireTokenHoldings(1)).to.not.be.reverted;
+            await expect(context.nested.nestedAssetBatcher.requireTokenHoldings(1)).to.not.be.reverted;
         });
     });
 });
