+++ test/unit/NestedBuyBacker.unit.ts
@@ -4,7 +4,7 @@ import { appendDecimals } from "../helpers";
 import { ethers } from "hardhat";
 import { expect } from "chai";
 import { DummyRouter, FeeSplitter, MockERC20, NestedBuybacker, WETH9 } from "../../typechain";
-import { describeWithoutFork } from "../shared/provider";
+import { describeWithoutFork } from "../shared/helpers/provider";
 
 describeWithoutFork("NestedBuybacker", () => {
     let alice: SignerWithAddress, bob: SignerWithAddress, communityReserve: SignerWithAddress;
